+++ frontend/src/components/Header/Notification/useNotification.ts
@@ -1,26 +1,12 @@
 import { useState } from 'react';
 import { useLocation } from 'react-router-dom';
-import { useRecoilValue } from 'recoil';
-import { isLoginState } from 'recoil/auth/atoms';
 
-import useSnackBar from 'hooks/useSnackBar';
-
-export const useHeaderRightButton = () => {
-  const isLogin = useRecoilValue(isLoginState);
+export const useNotification = () => {
   const [notificationCount, setNotificationCount] = useState<number | undefined>(
     undefined,
   );
   const [isSubscribed, setIsSubscribed] = useState(false);
-
   const { pathname } = useLocation();
-  const renderSnackbar = useSnackBar();
-
-  const handleClickErrorFallbackLoginButton = () => {
-    renderSnackbar({
-      status: 'ERROR',
-      message: '로그인 요청 시 에러가 발생했습니다. 새로 고침을 해주세요.',
-    });
-  };
 
   const updateNotificationCount = (updatedNotificationCount: number) => {
     setNotificationCount(updatedNotificationCount);
@@ -31,12 +17,10 @@ export const useHeaderRightButton = () => {
   };
 
   return {
-    isLogin,
-    pathname,
     notificationCount,
     isSubscribed,
-    handleClickErrorFallbackLoginButton,
     updateNotificationCount,
     updateIsSubscribed,
+    pathname,
   };
 };
