--- frontend/src/components/Header/useLoginButton.ts
@@ -1,13 +1,0 @@
-import useAuth from 'hooks/useAuth';
-
-export const useLoginButton = () => {
-  const { redirectGoogleLoginLink } = useAuth();
-
-  const handleClickLoginButton = () => {
-    redirectGoogleLoginLink();
-  };
-
-  return {
-    handleClickLoginButton,
-  };
-};
